requirements.txt
yfinance
vectorbt
pandas-ta
bokeh
streamlit
streamlit-aggrid
google-generativeai
requests